(clang): Add GameBoyTargetInfo to the Make list without failures.
GameBoyTargetInfo was not previously being built within Clang, which
meant that it was impossible to invoke Clang and use it as a target.
After adding it to the CMakeLists, some errors in the class were
corrected and the project now builds successfully. Clang can target
the Game Boy by providing the --target="gameboy" option.

N.B.: Clang will crash if given a C program for the time being. This
is because the back-end is still virtually empty.